Added support for tls-auth key
Added the necessary actions for enabling tls-auth with a simple
attribute. The new attribute is node["openvpn"]["tls-auth"] and takes a
relative path and a filename in which the TLS auth key will be saved.
The path to the file is anchored to whatever the
node["openvpn"]["key_dir"] is set to.

 * [new] add a new node["openvpn"]["tls-auth"] attribute to the
 cookbook. By default this attribute is unset. If tls-auth needs to be
 enabled, this attribute should be set to a filename. The filename
 will be created relative to the node["openvpn"]["key_dir"] path
 (by default /etc/openvpn/keys/)
 * [new] add support in the configuration templates for enabling tls-auth
 key directives.
 * [new] add a resource to the default.rb recipe to create a tls-auth
 preshared authorization key if it is requested